first run
successfully connect with the json-server and display the data inside the screen

+++ lib/models/semester.dart
@@ -22,8 +22,8 @@ class Semester {
       : this(
           semesterNo: json['semesterNo'],
           durationInWeek: json['durationInWeek'],
-          targetedGPA: json['targetedGPA'],
-          realGPA: json['realGPA'],
+          targetedGPA: json['targetedGPA'].toDouble(),
+          realGPA: json['realGPA'].toDouble(),
           totalCredit: json['totalCredit'],
           semesterStatus: json['semesterStatus'],
           courses: json['courses'] != null
